+++ src/main/scala/bitcoin/Codecs.scala
@@ -16,42 +16,12 @@
   */
 package io.tokenanalyst.bitcoinrpc.bitcoin
 
-import io.circe.{Encoder, Decoder, HCursor, Json}
-import io.tokenanalyst.bitcoinrpc.{RPCResponse, RPCRequest, RPCEncoder}
-import Protocol._
+import io.circe.{Encoder, Json}
+import io.tokenanalyst.bitcoinrpc.GenericRPCEncoders._
+import io.tokenanalyst.bitcoinrpc.bitcoin.Protocol._
+import io.tokenanalyst.bitcoinrpc.{RPCEncoder, RPCRequest}
 
-object RPCDecoders {
-  implicit def batch[A <: RPCResponse: Decoder] =
-    new Decoder[BatchResponse[A]] {
-      def apply(a: HCursor): Decoder.Result[BatchResponse[A]] =
-        a.as[Seq[A]].map(s => BatchResponse(s))
-    }
-
-  implicit def deriveCirceDecoder[A <: RPCResponse: Decoder] = new Decoder[A] {
-    def apply(a: HCursor): Decoder.Result[A] = a.downField("result").as[A]
-  }
-}
-
-object RPCEncoders {
-  def requestFields(
-      method: String,
-      params: Iterable[Json]
-  ): List[(String, Json)] = List(
-    ("jsonrpc", Json.fromString("2.0")),
-    ("id", Json.fromString("0")),
-    ("method", Json.fromString(method)),
-    ("params", Json.fromValues(params))
-  )
-
-  implicit def batchRequest[A <: RPCRequest](implicit encoder: RPCEncoder[A]) =
-    new RPCEncoder[BatchRequest[A]] {
-      final def apply(req: BatchRequest[A]): Json = {
-        val jsons = req.seq.map { i =>
-          encoder.apply(i)
-        }
-        Json.arr(jsons: _*)
-      }
-    }
+object Codecs {
 
   implicit val transactionRequest = new RPCEncoder[TransactionRequest] {
     final def apply(a: TransactionRequest): Json =
